Remove routing from Wyre KYC form
